pyproject.toml
[tool.poetry]
name = "py-aiger-improv"
version = "0.1.0"
description = "Create controlled improvisors to control aiger circuits."
authors = ["Marcell Vazquez-Chanlatte <mvc@linux.com>"]
license = "MIT"
packages = [
    { include = "aiger_improv"},
]

[tool.poetry.dependencies]
python = "^3.7"
mdd = "^0.3.4"
py-aiger-bdd = "^3.0.0"
py-aiger-bv = "^4.5.1"
attrs = "^20.2.0"
py-aiger = "^6.1.5"
py-aiger-discrete = "^0.1.6"
numpy = "^1.19.2"
py-aiger-coins = "^3.1.0"

[tool.poetry.dev-dependencies]
pytest = "^6.0.2"
pytest-flake8 = "^1.0.6"
py-aiger-ptltl = "^3.0.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
